I added the function "_init_argparse()" which takes arguments of generation, is net?, guess limit, and testing. First three are clear and the last one (testing) basically to check some things while developing the game. It gives you the answer at the start of the game end also increases the verbose.
I added variables "end" "tries" to the game object. Moreover added a new variable to the game called "won" for checking if the game is won (for celebration purposes).

I added _validator and _completer functions to the game object and change the input to inquirer input. With this, you can basically write a letter or half of the pokemon and press tab for autocompletion. If you click tab more, you will see that the options changes if there are more options matching. It requires inquirer package (exists in pip or conda-forge).

Added a game ending celebration called game_ending(self) function in the game object.

Small changes in the guess.